+++ library/build.gradle
@@ -1,14 +1,14 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.3"
+    compileSdkVersion 25
+    buildToolsVersion "25.0.2"
 
     defaultConfig {
-        minSdkVersion 7
-        targetSdkVersion 23
-        versionCode 10
-        versionName '1.5'
+        minSdkVersion 9
+        targetSdkVersion 25
+        versionCode 1
+        versionName '1.0'
     }
     buildTypes {
         release {
@@ -24,15 +24,15 @@ android {
 
 dependencies {
     compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile 'com.android.support:appcompat-v7:23.4.0'
-    compile 'com.android.support:preference-v7:23.4.0'
-    compile 'com.android.support:support-annotations:23.4.0'
+    compile 'com.android.support:appcompat-v7:25.1.1'
+    compile 'com.android.support:preference-v7:25.1.1'
+    compile 'com.android.support:support-annotations:25.1.1'
 }
 
 ext {
-    PUBLISH_GROUP_ID = 'com.pavelsikun'
-    PUBLISH_ARTIFACT_ID = 'vintage-chroma'
-    PUBLISH_VERSION = '1.5'
+    PUBLISH_GROUP_ID = 'com.kunzisoft'
+    PUBLISH_ARTIFACT_ID = 'androidclearchroma'
+    PUBLISH_VERSION = '1.0'
 }
 
 apply from: 'https://raw.githubusercontent.com/blundell/release-android-library/master/android-release-aar.gradle'
