+++ library/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest
-    package="com.pavelsikun.vintagechroma">
+    package="com.kunzisoft.androidclearchroma">
     <application/>
 </manifest>
