+++ library/src/main/java/com/kunzisoft/androidclearchroma/IndicatorMode.java
@@ -1,0 +1,8 @@
+package com.kunzisoft.androidclearchroma;
+
+/**
+ * Created by Pavel Sikun on 30.03.16.
+ */
+public enum IndicatorMode {
+    DECIMAL, HEX;
+}
