+++ library/src/main/java/com/kunzisoft/androidclearchroma/OnColorSelectedListener.java
@@ -1,0 +1,10 @@
+package com.kunzisoft.androidclearchroma;
+
+import android.support.annotation.ColorInt;
+
+/**
+ * Created by Pavel Sikun on 28.03.16.
+ */
+public interface OnColorSelectedListener {
+    void onColorSelected(@ColorInt int color);
+}
