+++ library/src/main/java/com/kunzisoft/androidclearchroma/ShapePreviewPreference.java
@@ -1,0 +1,10 @@
+package com.kunzisoft.androidclearchroma;
+
+/**
+ * Created by Jeremy JAMET on 12/09/16.
+ */
+public enum ShapePreviewPreference {
+    CIRCLE,
+    SQUARE,
+    ROUNDED_SQUARE
+}
