--- library/src/main/java/com/pavelsikun/vintagechroma/IndicatorMode.java
@@ -1,8 +1,0 @@
-package com.pavelsikun.vintagechroma;
-
-/**
- * Created by Pavel Sikun on 30.03.16.
- */
-public enum IndicatorMode {
-    DECIMAL, HEX;
-}
