--- library/src/main/java/com/pavelsikun/vintagechroma/OnColorSelectedListener.java
@@ -1,10 +1,0 @@
-package com.pavelsikun.vintagechroma;
-
-import android.support.annotation.ColorInt;
-
-/**
- * Created by Pavel Sikun on 28.03.16.
- */
-public interface OnColorSelectedListener {
-    void onColorSelected(@ColorInt int color);
-}
