--- library/src/main/java/com/pavelsikun/vintagechroma/ShapePreviewPreference.java
@@ -1,10 +1,0 @@
-package com.pavelsikun.vintagechroma;
-
-/**
- * Created by Jeremy JAMET on 12/09/16.
- */
-public enum ShapePreviewPreference {
-    CIRCLE,
-    SQUARE,
-    ROUNDED_SQUARE
-}
