+++ sample/build.gradle
@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.3"
+    compileSdkVersion 25
+    buildToolsVersion "25.0.2"
 
     defaultConfig {
-    applicationId "com.pavelsikun.vintagechroma.sample"
-        minSdkVersion 7
-        targetSdkVersion 23
+    applicationId "com.kunzisoft.androidclearchroma.sample"
+        minSdkVersion 9
+        targetSdkVersion 25
         versionCode 1
         versionName "1.0"
     }
@@ -27,12 +27,9 @@ dependencies {
     testCompile 'junit:junit:4.12'
     compile project(':library')
 
-//    for your projects use:
-//    compile 'com.pavelsikun:vintage-chroma:1.5'
-
-    compile 'com.android.support:cardview-v7:23.4.0'
-    compile 'com.android.support:appcompat-v7:23.4.0'
-    compile 'com.android.support:design:23.4.0'
-    compile 'com.android.support:support-annotations:23.4.0'
-    compile 'com.android.support:preference-v7:23.4.0'
+    compile 'com.android.support:cardview-v7:25.1.1'
+    compile 'com.android.support:appcompat-v7:25.1.1'
+    compile 'com.android.support:design:25.1.1'
+    compile 'com.android.support:support-annotations:25.1.1'
+    compile 'com.android.support:preference-v7:25.1.1'
 }
