+++ sample/src/main/java/com/kunzisoft/androidclearchroma/sample/PreferencesActivity.java
@@ -1,0 +1,46 @@
+package com.kunzisoft.androidclearchroma.sample;
+
+import android.annotation.TargetApi;
+import android.os.Build;
+import android.os.Bundle;
+import android.preference.PreferenceFragment;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.AppCompatActivity;
+import android.view.MenuItem;
+
+/**
+ * Created by Pavel Sikun on 31.03.16.
+ */
+
+@TargetApi(Build.VERSION_CODES.HONEYCOMB)
+public class PreferencesActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        getFragmentManager()
+                .beginTransaction()
+                .replace(android.R.id.content, new PreferencesScreen())
+                .commit();
+
+        ActionBar toolbar = getSupportActionBar();
+        if(toolbar != null) {
+            toolbar.setDisplayHomeAsUpEnabled(true);
+        }
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if(item.getItemId() == android.R.id.home) {
+            onBackPressed();
+        }
+        return super.onOptionsItemSelected(item);
+    }
+
+    public static class PreferencesScreen extends PreferenceFragment {
+        public void onCreate(Bundle savedInstanceState) {
+            super.onCreate(savedInstanceState);
+            addPreferencesFromResource(R.xml.pref);
+        }
+    }
+}
