+++ sample/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.pavelsikun.vintagechroma.sample.MainActivity">
+    tools:context="com.kunzisoft.androidclearchroma.sample.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
