+++ sample/src/main/res/xml/pref.xml
@@ -2,20 +2,20 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <com.pavelsikun.vintagechroma.ChromaPreference
+    <com.kunzisoft.androidclearchroma.ChromaPreference
         android:key="rgb"
         android:title="RGB sample"
         android:summary="[color]"
         android:enabled="false"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreference
+    <com.kunzisoft.androidclearchroma.ChromaPreference
         android:key="argb"
         android:title="ARGB sample"
         android:summary="Example [color]"
         app:chromaColorMode="ARGB"
         app:chromaInitialColor="#00ffff"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreference
+    <com.kunzisoft.androidclearchroma.ChromaPreference
         android:key="argbhex"
         android:title="ARGB and HEX indicator sample"
         android:summary="[color]"
@@ -23,13 +23,13 @@
         app:chromaIndicatorMode="HEX"
         app:chromaInitialColor="@color/colorPrimary"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreference
+    <com.kunzisoft.androidclearchroma.ChromaPreference
         android:key="hsv"
         android:title="HSV sample without summary"
         app:chromaColorMode="HSV"
         app:chromaInitialColor="@color/colorAccent"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreference
+    <com.kunzisoft.androidclearchroma.ChromaPreference
         android:key="roundedquare"
         android:title="ARGB sample with rounded square sample"
         android:summary="Sample [color] summary"
@@ -37,7 +37,7 @@
         app:chromaColorMode="ARGB"
         app:chromaInitialColor="#ffee00"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreference
+    <com.kunzisoft.androidclearchroma.ChromaPreference
         android:key="customsummary"
         android:title="ARGB with square and simple summary"
         android:summary="Sample summary"
