+++ sample/src/main/res/xml/prefs_v7.xml
@@ -2,40 +2,40 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <com.pavelsikun.vintagechroma.ChromaPreferenceCompat
+    <com.kunzisoft.androidclearchroma.ChromaPreferenceCompat
         android:key="rgb"
         android:title="RGB sample"
         android:summary="[color]"
         android:enabled="false"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreferenceCompat
+    <com.kunzisoft.androidclearchroma.ChromaPreferenceCompat
         android:key="argb"
         android:title="ARGB sample"
         android:summary="[color]"
         app:chromaColorMode="ARGB"
         app:chromaInitialColor="@android:color/black"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreferenceCompat
+    <com.kunzisoft.androidclearchroma.ChromaPreferenceCompat
         android:key="argbhex"
         android:title="ARGB and HEX indicator sample"
         app:chromaColorMode="ARGB"
         app:chromaIndicatorMode="HEX"
         app:chromaInitialColor="@color/colorPrimary"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreferenceCompat
+    <com.kunzisoft.androidclearchroma.ChromaPreferenceCompat
         android:key="hsv"
         android:title="HSV sample"
         app:chromaColorMode="HSV"
         app:chromaInitialColor="@color/colorAccent"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreferenceCompat
+    <com.kunzisoft.androidclearchroma.ChromaPreferenceCompat
         android:key="roundedquare"
         android:title="ARGB sample with rounded square sample"
         app:chromaShapePreview="ROUNDED_SQUARE"
         app:chromaColorMode="ARGB"
         app:chromaInitialColor="#ffee00"/>
 
-    <com.pavelsikun.vintagechroma.ChromaPreferenceCompat
+    <com.kunzisoft.androidclearchroma.ChromaPreferenceCompat
         android:key="customsummary"
         android:title="ARGB sample with square sample and custom summary"
         android:summary="Example summary"
